+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### Features
 
 -   **terra-form**
-    -   **text-area** replaced deprecated `terra-text-area-input` with a `mat-input` variant.
+    -   form type **number**: replaced deprecated `terra-number-input` with a `mat-input` variant.
+    -   form type **text-area** replaced deprecated `terra-text-area-input` with a `mat-input` variant.
 
 # 11.1.0 (23.03.2021)